bot/requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
googlesearch-python==1.2.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.7.1
pydantic_core==2.18.2
python-telegram-bot==21.1.1
pytube==15.0.0
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
youtube-search-python==1.6.6
